refactor: extract wipe_controller.py and timeline_widgets.py
Split ui/widgets/split_view.py (822 -> 721 lines):
- wipe_controller.py: wipe geometry, painting, drag handling functions

Split ui/widgets/frame_scrubber.py (715 -> 375 lines):
- timeline_widgets.py: _FatSlider, CoverageBar, MarkerOverlay widgets

All 427 tests pass.
